Revert "Syncing, fixing bugs/polish"
This reverts commit c628ac29a78297cf975b1335a239936fedba6b42.

+++ Assets/Scripts/Iinteraction.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Iinteraction : MonoBehaviour
 {
-    [SerializeField] float speedBoost;
+    [SerializeField] int speedBoost;
     [SerializeField] int jumpCIncrease;
     [SerializeField] int damageBoostFlat;
     [SerializeField] float damageBoostMult;
